fix: 修复 4 个 clippy 错误 + 6 个相关衍生错误（cargo clippy --all-targets -- -D warnings 通过）
原报：
- trajectory/rl.rs:543/607 `clippy::unnecessary_cast`：`r.step_index as usize` 多余（已为 usize）
- agent/provider_fallback.rs:274 `clippy::unnecessary_map_or`：`map_or(false, ...)` 应改 `is_some_and`
- agent/session_manager.rs:406 `clippy::clone_on_copy`：`m.usage.clone()` 多余（Option<TokenUsage> 是 Copy）

衍生修复（clippy --all-targets -- -D warnings 暴露）：
- gateway/native.rs:1285/1709 + gateway/routes.rs:272：GatewayAppState 缺 client_ip_policy 字段（之前 GatewayAppState 加字段时未同步测试 init）
- gateway/routes.rs：把 ClientIpPolicy 从 crate 顶层 use 移到 #[cfg(test)] mod tests 内的 use（避免 lib 编译 unused import 警告）
- gateway/auth.rs:65 `clippy::manual_contains`：`iter().any(|x| ip == *x)` 改 `contains(&ip)`
- orchestrator/executor.rs:738 `unused_variables`：`for round in 0..=...` 改 `for _round in ...`
- providers/openclaw.rs:27 `clippy::needless_lifetimes`：`<'a>` 多余，删除
- rt-workflow/dispatcher.rs:41 `clippy::let_and_return`：`let dispatcher = Self {...}; dispatcher` 改为直接返回 `Self {...}`
- commands/memory.rs:917/1189 `clippy::let_underscore_future`：`let _ = future` 会丢弃 future 而不 poll，导致 async 操作永不执行；改用 `if let Err(e) = future.await` / `let r = future.await; if !r.success` 模式

验证：cargo clippy --all-targets -- -D warnings exit 0、cargo fmt 0 diff、tsc --noEmit 0 error

+++ src-tauri/crates/trajectory/src/rl.rs
@@ -540,7 +540,7 @@ impl RLEngine {
         // 避免索引错配（原代码直接用 rewards[t] 当作 step t 的奖励，但 rewards 顺序与 step 顺序未必一致）
         let mut step_rewards: Vec<f64> = vec![0.0; values.len()];
         for r in rewards {
-            let idx = r.step_index as usize;
+            let idx = r.step_index;
             if idx < step_rewards.len() {
                 step_rewards[idx] += r.value;
             }
@@ -604,7 +604,7 @@ impl RLEngine {
         // 先按 step_index 聚合 rewards → step_rewards
         let mut step_rewards: Vec<f64> = vec![0.0; steps];
         for r in &trajectory.rewards {
-            let idx = r.step_index as usize;
+            let idx = r.step_index;
             if idx < step_rewards.len() {
                 step_rewards[idx] += r.value;
             }
